Replace walker.rs with new_walker.rs.
